Fix for inability to provide custom implementation of AsyncClientHttpRequestFactory
